requirements.txt
aiogram==3.6.0
edge-tts==6.1.12
python-dotenv==1.0.1
